Major fixes and Matchmaking Implementation
Lacks the queue part

+++ GeneralLordV0.1/Client/Model/ArmyContainer.cs
@@ -10,7 +10,7 @@ namespace GeneralLordWebApiClient.Model
     [Serializable]
     public class ArmyContainer
     {
-        public int ArmyStrength;
+        public float ArmyStrength;
         public List<TroopContainer> TroopContainers { get; set; }
 
         public string CharacterXML { get; set; }
